Renaming projects to definitions.
