+++ plasoscaffolder/definitions/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+"""This file imports Python modules that registers definitions."""
+
+from plasoscaffolder.definitions import plaso
