--- plasoscaffolder/projects/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-"""This file imports Python modules that registers projects."""
-
-from plasoscaffolder.projects import plaso
