+++ run_tests.py
@@ -10,8 +10,8 @@ if __name__ == '__main__':
       'tests.common', pattern='*.py')
   test_suite_bll = unittest.TestLoader().discover('tests.bll', pattern='*.py')
   test_suite_dal = unittest.TestLoader().discover('tests.dal', pattern='*.py')
-  test_suite_projects = unittest.TestLoader().discover(
-      'tests.projects', pattern='*.py')
+  test_suite_definitions = unittest.TestLoader().discover(
+      'tests.definitions', pattern='*.py')
   test_suite_frontend = unittest.TestLoader().discover(
       'tests.frontend', pattern='*.py')
   test_suite_model = unittest.TestLoader().discover(
@@ -23,7 +23,7 @@ if __name__ == '__main__':
   all_tests = unittest.TestSuite((test_suite_bll, test_suite_common,
                                   test_suite_dal, test_suite_frontend,
                                   test_suite_model, test_suite_other,
-                                  test_suite_projects,
+                                  test_suite_definitions,
                                   test_suite_integrationtests))
 
   test_results = unittest.TextTestRunner(verbosity=2).run(all_tests)
