+++ tests/definitions/PlasoFailPath1/__init__.py
