+++ tests/definitions/PlasoFailPath1/foobar/__init__.py
