+++ tests/definitions/PlasoFailPath1/plaso/__init__.py
