+++ tests/definitions/PlasoFailPath2/plaso/__init__.py
