+++ tests/definitions/PlasoFailPath2/plaso/parsers/__init__.py
