+++ tests/definitions/PlasoPath/plaso/__init__.py
