+++ tests/definitions/PlasoPath/plaso/parsers/__init__.py
